Add shell commands

+++ irbrc
@@ -1,9 +1,47 @@
 require 'irb/completion'
 
-IRB.conf[:SAVE_HISTORY] = 1000
 IRB.conf[:AUTO_INDENT] = true
+IRB.conf[:SAVE_HISTORY] = 1000
 IRB.conf[:PROMPT_MODE] = :SIMPLE
 
+def tmpfile
+  '_edit_irb_tmp_file.rb'
+end
+
+
+def edit
+  system("vim #{tmpfile}")
+  eval File.read(tmpfile)
+ensure
+  # File.delete(tmpfile)
+end
+
+def ri(method)
+  puts `ri #{method}`
+end
+
+def vim(filename)
+  system("vim #{filename}")
+end
+
+# shell
+def ls(path = nil)
+  if path
+    Dir["#{path}/*"]
+  else
+    Dir["*"]
+  end
+end
+
+def cd(path)
+  Dir.chdir(path)
+end
+
+def pwd
+  Dir.pwd
+end
+
+# benchmark 
 def time
   start = Time.now
   yield
@@ -11,10 +49,7 @@ ensure
   puts "Time: #{Time.now - start} seconds"
 end
 
-def edit
-  tmpfile = '_edit_irb_tmp_file.rb'
-  system("vim #{tmpfile}")
-  eval File.read(tmpfile)
-ensure
-  File.delete(tmpfile)
+at_exit do
+  File.delete(tmpfile) if File.exists?(tmpfile)
 end
+
